software_developer/requirements.txt
uagents>=0.6.0
uagents-core>=0.2.0
fastmcp>=0.1.0
pydantic>=2.0.0
aiohttp>=3.8.5
python-dotenv>=1.0.0